+++ project/project-app/src/main/java/lt/baltupiusoftas/project/app/HeaderStatusBean.java
@@ -1,14 +1,6 @@
 package lt.baltupiusoftas.project.app;
 
-import lt.baltupiusoftas.project.domain.User;
-import lt.baltupiusoftas.project.service.password.PasswordHashingService;
-import lt.baltupiusoftas.project.service.user.UserService;
-
-import javax.enterprise.context.RequestScoped;
 import javax.enterprise.context.SessionScoped;
-import javax.faces.application.FacesMessage;
-import javax.faces.context.FacesContext;
-import javax.inject.Inject;
 import javax.inject.Named;
 import java.io.Serializable;
