+++ project/project-app/src/main/java/lt/baltupiusoftas/project/app/user/UserLoginBean.java
@@ -4,8 +4,8 @@ package lt.baltupiusoftas.project.app.user;
 import lt.baltupiusoftas.project.app.HeaderStatusBean;
 import lt.baltupiusoftas.project.app.Login;
 import lt.baltupiusoftas.project.domain.User;
-import lt.baltupiusoftas.project.service.password.PasswordHashingService;
-import lt.baltupiusoftas.project.service.user.UserService;
+import lt.baltupiusoftas.project.service.PasswordHashingService;
+import lt.baltupiusoftas.project.service.UserService;
 
 import javax.enterprise.context.RequestScoped;
 import javax.faces.application.FacesMessage;
@@ -46,9 +46,16 @@ public class UserLoginBean implements Serializable {
         User user = userService.login(email, password);
         // If user is registered and login successful
         if (user != null) {
-            login.setUser(user);
+            if (user.getBlocked())
+            {
+                FacesContext.getCurrentInstance().addMessage("loginBtn", new FacesMessage(FacesMessage.SEVERITY_ERROR,"Klaida!",  "Vartotojas užblokuotas. Jei manote, kad įvyko klaida, kreipkitės į sistemos administratorių."));
+                logout();
+            }
+            else {
+                login.setUser(user);
 
-            headerStatusBean.showLogoutAndUserProfile();
+                headerStatusBean.showLogoutAndUserProfile();
+            }
             return "index";
         }
         // If user is not registered and login failed
@@ -71,7 +78,7 @@ public class UserLoginBean implements Serializable {
 
 
     private Boolean isLoggedIn() {
-        return login.getUser().getEmail() != null;
+        return login.getUser() != null;
     }
 
     public String getEmail() {
